Re-organized styles. Fixed up the details page. Stylized new update page
