--- client/src/assets/styles/components/order-list.css
@@ -1,32 +1,0 @@
-body{
-    background-color: #F4F0E5;
-}
-
-.header-container{
-    @apply flex items-center px-20 py-5 sticky top-0;
-    background-color: #F4F0E5;
-    border-bottom: 1px solid black;
-}
-
-.main-container{
-    @apply px-20 py-5 pb-0;
-}
-
-.profile-icon{
-    @apply flex ml-auto items-center;
-}
-
-.profile-name{
-    @apply my-0 mx-2.5;
-}
-
-.pagination-container{
-    @apply float-right;
-    padding: 0 5%;
-}
-
-.counter{
-    @apply font-semibold my-2.5 mx-0;
-    color: #ff642e;
-    margin-right:5%;
-}
