+++ client/src/containers/order/DeliveryTypeListContainer.tsx
@@ -2,7 +2,7 @@ import { useDispatch } from "react-redux"
 import { AppDispatch } from "../../state/redux"
 import { useNavigate } from "react-router-dom"
 import { useCallback, useEffect, useMemo, useState } from "react"
-import DeliveryTypeList from "../../components/order/DeliveryList"
+import DeliveryTypeList from "../../components/order/admin/DeliveryList"
 import { DeliverySort, DeliveryType } from "../../types"
 import { fetchDeliveryTypesService } from "../../services"
 import { addNotification } from "../../state/redux/slices/notificationSlice"
